requirements.txt
requests==2.31.0
python-dotenv==1.0.0
py3cw==0.0.36
pycoingecko==3.1.0
schedule==1.2.0
pandas
pyyaml==6.0.1
loguru==0.7.0
websockets==11.0.3
sqlalchemy==2.0.23
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
PyYAML==6.0.1
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
alembic==1.12.1
